Newlines at eof, rename util
