+++ library/base/manager.js
