+++ library/ift.js
@@ -52,4 +52,5 @@ var ift = module.exports = {
     mixin: require('./util/mixin'),
     debug: require('./util/debug')
   }
+
 };
