+++ library/util/bind.js
